+++ R/saprocessing.R
@@ -305,7 +305,7 @@ transfer_sa_item <- function(jsap_from, jsap_to, selected_sa_items,
         }
 
         if (print_indications) {
-             cat("Serie ", serie_name, ": Transfered.\n")
+            cat("Serie ", serie_name, ": Transfered.\n")
         }
     }
 
@@ -635,11 +635,15 @@ set_name <- function(jsap, idx, name) {
     replace_sa_item(jsap, jsai = jsai, idx = idx)
 }
 
-#' @title Set (JDemetra+) Time Series Metadata of a SA-item
+#' @title Set (JDemetra+) Metadata of a SA-item
 #'
 #' @description
-#' Function to set the time series metadata of a SA-item (provider, source of the data...).
-#' `set_ts_metadata()` uses the metadata of another SA-item while `put_ts_metadata()`
+#' Function to set the metadata of a SA-item.
+#'
+#' `XXX_ts_metadata()` set the time series metadata of a SA-item (provider,
+#' source of the data...). `XXX_metadata()` set any metadata to a SA-Item.
+#'
+#' `set_XXX()` uses the metadata of another SA-item while `put_XXX()`
 #' allows to update a specific key with a new information.
 #'
 #' @inheritParams set_raw_data
@@ -661,41 +665,6 @@ set_name <- function(jsap, idx, name) {
 #' jsai <- jsap_sai(jsap, 1)
 #' get_ts_metadata(jsai, "@id")
 #'
-
-
-# set_metadata <- function(jsap, ref_jsai, idx) {
-#   jsai <- jsap_sai(jsap, idx = idx)
-#   jsai <- jsa$withInformations(ref_jsai$getMeta())
-#   replace_sa_item(jsap, jsai = jsai, idx = idx)
-# }
-
-# put_metadata <- function (jsap, idx, key, value) {
-#     jsai <- jsap_sai(jsap, idx = idx)
-#
-#     meta <- .jcall(jsai, "Ljava/util/Map;", "getMeta")
-#     new_meta <- .jnew("java/util/HashMap", meta)
-#
-#     jkey <- .jnew("java/lang/String", key)
-#     jvalue <- .jnew("java/lang/String", value)
-#
-#     .jcall(
-#         obj = new_meta,
-#         returnSig = "Ljava/lang/Object;",
-#         method = "put",
-#         .jcast(jkey, "java/lang/Object"),
-#         .jcast(jvalue, "java/lang/Object")
-#     )
-#
-#     jsai <- .jcall(
-#         obj = jsai,
-#         returnSig = "Ljdplus/sa/base/api/SaItem;",
-#         method = "withInformations",
-#         .jcast(new_meta, "java/util/Map")
-#     )
-#
-#     replace_sa_item(jsap, jsai = jsai, idx = idx)
-# }
-
 set_ts_metadata <- function(jsap, idx, ref_jsai) {
 
     jsai <- jsap_sai(jsap, idx = idx)
@@ -739,6 +708,45 @@ put_ts_metadata <- function(jsap, idx, key, value) {
     )
     replace_sa_item(jsap, jsai = jsai, idx = idx)
 }
+
+
+#' @name set_ts_metadata
+#' @export
+set_metadata <- function(jsap, ref_jsai, idx) {
+    jsai <- jsap_sai(jsap, idx = idx)
+    jsai <- jsa$withInformations(ref_jsai$getMeta())
+    replace_sa_item(jsap, jsai = jsai, idx = idx)
+}
+
+#' @name set_ts_metadata
+#' @export
+put_metadata <- function (jsap, idx, key, value) {
+    jsai <- jsap_sai(jsap, idx = idx)
+
+    meta <- .jcall(jsai, "Ljava/util/Map;", "getMeta")
+    new_meta <- .jnew("java/util/HashMap", meta)
+
+    jkey <- .jnew("java/lang/String", key)
+    jvalue <- .jnew("java/lang/String", value)
+
+    .jcall(
+        obj = new_meta,
+        returnSig = "Ljava/lang/Object;",
+        method = "put",
+        .jcast(jkey, "java/lang/Object"),
+        .jcast(jvalue, "java/lang/Object")
+    )
+
+    jsai <- .jcall(
+        obj = jsai,
+        returnSig = "Ljdplus/sa/base/api/SaItem;",
+        method = "withInformations",
+        .jcast(new_meta, "java/util/Map")
+    )
+
+    replace_sa_item(jsap, jsai = jsai, idx = idx)
+}
+
 
 
 #' @title Get/Set SA-item Priority